+++ CHANGES.md
@@ -1,5 +1,10 @@
 <!-- vim:ts=4:sts=4:sw=4:et:tw=60 -->
 
+## Next minor
+
+ - Deprecated ``ref`` in favor of ``resolve`` as recommended by
+   @domenicdenicola.
+
 ## 0.8.1
 
  - Fixed opera bug. #35 @cadorn